+++ openmdao/recorders/tests/test_sqlite_reader.py
@@ -1,6 +1,5 @@
 """ Unit tests for the SqliteCaseReader. """
 
-import errno
 import sys
 import os
 import sys
@@ -16,13 +15,11 @@ import openmdao.api as om
 from openmdao import __version__ as openmdao_version
 from openmdao.recorders.sqlite_recorder import format_version
 from openmdao.recorders.sqlite_reader import SqliteCaseReader
-from openmdao.recorders.tests.test_sqlite_recorder import ParaboloidProblem
 from openmdao.recorders.case import PromAbsDict
+from openmdao.recorders.tests.recorder_test_utils import assert_model_matches_case
 from openmdao.core.tests.test_discrete import ModCompEx, ModCompIm
 from openmdao.core.tests.test_expl_comp import RectangleComp, RectangleCompWithTags
-from openmdao.core.tests.test_units import SpeedComp
 from openmdao.test_suite.components.eggcrate import EggCrate
-from openmdao.test_suite.components.expl_comp_array import TestExplCompArray
 from openmdao.test_suite.components.implicit_newton_linesearch import ImplCompTwoStates
 from openmdao.test_suite.components.paraboloid import Paraboloid
 from openmdao.test_suite.components.paraboloid_problem import ParaboloidProblem
@@ -35,19 +32,11 @@ from openmdao.utils.assert_utils import assert_near_equal, assert_warning
 from openmdao.utils.general_utils import set_pyoptsparse_opt, determine_adder_scaler, printoptions
 from openmdao.utils.general_utils import remove_whitespace
 from openmdao.utils.testing_utils import use_tempdirs
-from openmdao.utils.mpi import MPI, multi_proc_exception_check
-from openmdao.utils.units import convert_units
-from openmdao.utils.om_warnings import OMDeprecationWarning
 
 # check that pyoptsparse is installed
 OPT, OPTIMIZER = set_pyoptsparse_opt('SLSQP')
 if OPTIMIZER:
     from openmdao.drivers.pyoptsparse_driver import pyOptSparseDriver
-
-try:
-    from openmdao.vectors.petsc_vector import PETScVector
-except ImportError:
-    PETScVector = None
 
 
 def count_keys(d):
@@ -1950,284 +1939,6 @@ class TestSqliteCaseReader(unittest.TestCase):
             for k in expected_set:
                 np.testing.assert_almost_equal(expected_set[k], actual_set[k])
 
-    def test_simple_load_system_cases(self):
-        prob = SellarProblem()
-
-        model = prob.model
-        model.recording_options['record_inputs'] = True
-        model.recording_options['record_outputs'] = True
-        model.recording_options['record_residuals'] = True
-        model.add_recorder(self.recorder)
-
-        prob.setup()
-
-        prob.run_driver()
-        prob.cleanup()
-
-        cr = om.CaseReader(self.filename)
-
-        system_cases = cr.list_cases('root', out_stream=None)
-        case = cr.get_case(system_cases[0])
-
-        # Add one to all the inputs and outputs just to change the model
-        #   so we can see if loading the case values really changes the model
-        for name in model._inputs:
-            model._inputs[name] += 1.0
-        for name in model._outputs:
-            model._outputs[name] += 1.0
-
-        # Now load in the case we recorded
-        prob.load_case(case)
-
-        _assert_model_matches_case(case, model)
-
-    def test_load_bad_system_case(self):
-        prob = SellarProblem(SellarDerivativesGrouped)
-
-        prob.model.add_recorder(self.recorder)
-
-        driver = prob.driver = om.ScipyOptimizeDriver()
-        driver.options['optimizer'] = 'SLSQP'
-        driver.options['tol'] = 1e-9
-        driver.options['disp'] = False
-        driver.recording_options['record_desvars'] = True
-        driver.recording_options['record_objectives'] = True
-        driver.recording_options['record_constraints'] = True
-
-        prob.setup()
-        prob.run_driver()
-        prob.cleanup()
-
-        cr = om.CaseReader(self.filename)
-
-        system_cases = cr.list_cases('root', out_stream=None)
-        case = cr.get_case(system_cases[0])
-
-        # try to load it into a completely different model
-        prob = SellarProblem()
-        prob.setup()
-
-        error_msg = "Input variable, '[^']+', recorded in the case is not found in the model"
-        with self.assertRaisesRegex(KeyError, error_msg):
-            prob.load_case(case)
-
-    def test_subsystem_load_system_cases(self):
-        prob = SellarProblem()
-        prob.setup()
-
-        model = prob.model
-        model.recording_options['record_inputs'] = True
-        model.recording_options['record_outputs'] = True
-        model.recording_options['record_residuals'] = True
-
-        # Only record a subsystem
-        model.d2.add_recorder(self.recorder)
-
-        prob.run_driver()
-        prob.cleanup()
-
-        cr = om.CaseReader(self.filename)
-
-        system_cases = cr.list_cases('root.d2', out_stream=None)
-        case = cr.get_case(system_cases[0])
-
-        # Add one to all the inputs just to change the model
-        #   so we can see if loading the case values really changes the model
-        for name in prob.model._inputs:
-            model._inputs[name] += 1.0
-        for name in prob.model._outputs:
-            model._outputs[name] += 1.0
-
-        # Now load in the case we recorded
-        prob.load_case(case)
-
-        _assert_model_matches_case(case, model.d2)
-
-    def test_load_system_cases_with_units(self):
-        comp = om.IndepVarComp()
-        comp.add_output('distance', val=1., units='m')
-        comp.add_output('time', val=1., units='s')
-
-        prob = om.Problem()
-        model = prob.model
-        model.add_subsystem('c1', comp)
-        model.add_subsystem('c2', SpeedComp())
-        model.add_subsystem('c3', om.ExecComp('f=speed', speed={'units': 'm/s'}, f={'units': 'm/s'}))
-        model.connect('c1.distance', 'c2.distance')
-        model.connect('c1.time', 'c2.time')
-        model.connect('c2.speed', 'c3.speed')
-
-        model.add_recorder(self.recorder)
-
-        prob.setup()
-        prob.run_model()
-
-        cr = om.CaseReader(self.filename)
-
-        system_cases = cr.list_cases('root', out_stream=None)
-        case = cr.get_case(system_cases[0])
-
-        # Add one to all the inputs just to change the model
-        # so we can see if loading the case values really changes the model
-        for name in model._inputs:
-            model._inputs[name] += 1.0
-        for name in model._outputs:
-            model._outputs[name] += 1.0
-
-        # Now load in the case we recorded
-        prob.load_case(case)
-
-        _assert_model_matches_case(case, model)
-
-        # make sure it still runs with loaded values
-        prob.run_model()
-
-        # make sure the loaded unit strings are compatible with `convert_units`
-        outputs = case.list_outputs(explicit=True, implicit=True, val=True,
-                                    units=True, shape=True, out_stream=None)
-        meta = {}
-        for name, vals in outputs:
-            meta[name] = vals
-
-        from_units = meta['c2.speed']['units']
-        to_units = meta['c3.f']['units']
-
-        self.assertEqual(from_units, 'km/h')
-        self.assertEqual(to_units, 'm/s')
-
-        self.assertEqual(convert_units(10., from_units, to_units), 10000./3600.)
-
-    def test_optimization_load_system_cases(self):
-        prob = SellarProblem(SellarDerivativesGrouped, nonlinear_solver=om.NonlinearRunOnce,
-                                                       linear_solver=om.ScipyKrylov,
-                                                       mda_nonlinear_solver=om.NonlinearBlockGS)
-
-        prob.model.add_recorder(self.recorder)
-
-        driver = prob.driver = om.ScipyOptimizeDriver()
-        driver.options['optimizer'] = 'SLSQP'
-        driver.options['tol'] = 1e-9
-        driver.options['disp'] = False
-        driver.recording_options['record_desvars'] = True
-        driver.recording_options['record_objectives'] = True
-        driver.recording_options['record_constraints'] = True
-
-        prob.setup()
-        prob.run_driver()
-        prob.cleanup()
-
-        inputs_before = prob.model.list_inputs(val=True, units=True, out_stream=None)
-        outputs_before = prob.model.list_outputs(val=True, units=True, out_stream=None)
-
-        cr = om.CaseReader(self.filename)
-
-        # get third case
-        system_cases = cr.list_cases('root', out_stream=None)
-        third_case = cr.get_case(system_cases[2])
-
-        iter_count_before = driver.iter_count
-
-        # run the model again with a fresh model
-        prob = SellarProblem(SellarDerivativesGrouped, nonlinear_solver=om.NonlinearRunOnce,
-                                                       linear_solver=om.ScipyKrylov,
-                                                       mda_nonlinear_solver=om.NonlinearBlockGS)
-
-        driver = prob.driver = om.ScipyOptimizeDriver()
-        driver.options['optimizer'] = 'SLSQP'
-        driver.options['tol'] = 1e-9
-        driver.options['disp'] = False
-
-        prob.setup()
-        prob.load_case(third_case)
-        prob.run_driver()
-        prob.cleanup()
-
-        inputs_after = prob.model.list_inputs(val=True, units=True, out_stream=None)
-        outputs_after = prob.model.list_outputs(val=True, units=True, out_stream=None)
-
-        iter_count_after = driver.iter_count
-
-        for before, after in zip(inputs_before, inputs_after):
-            np.testing.assert_almost_equal(before[1]['val'], after[1]['val'])
-
-        for before, after in zip(outputs_before, outputs_after):
-            np.testing.assert_almost_equal(before[1]['val'], after[1]['val'])
-
-        # Should take one less iteration since we gave it a head start in the second run
-        self.assertEqual(iter_count_before, iter_count_after + 1)
-
-    def test_load_solver_cases(self):
-        prob = SellarProblem()
-        prob.setup()
-
-        model = prob.model
-        model.nonlinear_solver.add_recorder(self.recorder)
-
-        fail = prob.run_driver()
-        prob.cleanup()
-
-        self.assertFalse(fail, 'Problem failed to converge')
-
-        cr = om.CaseReader(self.filename)
-
-        solver_cases = cr.list_cases('root.nonlinear_solver', out_stream=None)
-        case = cr.get_case(solver_cases[0])
-
-        # Add one to all the inputs just to change the model
-        # so we can see if loading the case values really changes the model
-        for name in prob.model._inputs:
-            model._inputs[name] += 1.0
-        for name in prob.model._outputs:
-            model._outputs[name] += 1.0
-
-        # Now load in the case we recorded
-        prob.load_case(case)
-
-        _assert_model_matches_case(case, model)
-
-    def test_load_driver_cases(self):
-        prob = om.Problem()
-        model = prob.model
-
-        model.add_subsystem('p1', om.IndepVarComp('x', 50.0), promotes=['*'])
-        model.add_subsystem('p2', om.IndepVarComp('y', 50.0), promotes=['*'])
-        model.add_subsystem('comp', Paraboloid(), promotes=['*'])
-        model.add_subsystem('con', om.ExecComp('c = x - y'), promotes=['*'])
-
-        model.add_design_var('x', lower=-50.0, upper=50.0)
-        model.add_design_var('y', lower=-50.0, upper=50.0)
-
-        model.add_objective('f_xy')
-        model.add_constraint('c', lower=15.0)
-
-        prob.driver.add_recorder(self.recorder)
-        prob.driver.recording_options['includes'] = ['*']
-
-        prob.set_solver_print(0)
-
-        prob.setup()
-        fail = prob.run_driver()
-        prob.cleanup()
-
-        self.assertFalse(fail, 'Problem failed to converge')
-
-        cr = om.CaseReader(self.filename)
-
-        driver_cases = cr.list_cases('driver', out_stream=None)
-        case = cr.get_case(driver_cases[0])
-
-        # Add one to all the inputs just to change the model
-        #   so we can see if loading the case values really changes the model
-        for name in prob.model._inputs:
-            prob.model._inputs[name] += 1.0
-        for name in prob.model._outputs:
-            prob.model._outputs[name] += 1.0
-
-        # Now load in the case we recorded
-        prob.load_case(case)
-
-        _assert_model_matches_case(case, model)
-
     def test_system_options_pickle_fail(self):
         # simple paraboloid model
         model = om.Group()
@@ -2387,145 +2098,6 @@ class TestSqliteCaseReader(unittest.TestCase):
             for key in case_type.list_cases():
                 self.assertTrue(key in case_type._cases)
                 self.assertEqual(key, case_type._cases[key].name)
-
-    def test_reading_driver_cases_with_indices(self):
-        # note: size must be an even number
-        SIZE = 10
-        prob = om.Problem()
-
-        driver = prob.driver = om.ScipyOptimizeDriver(disp=False)
-
-        prob.driver.add_recorder(self.recorder)
-        driver.recording_options['includes'] = ['*']
-
-        model = prob.model
-        indeps = model.add_subsystem('indeps', om.IndepVarComp(), promotes_outputs=['*'])
-
-        # the following were randomly generated using np.random.random(10)*2-1 to randomly
-        # disperse them within a unit circle centered at the origin.
-        # Also converted this array to > 1D array to test that capability of case recording
-        x_vals = np.array([
-            0.55994437, -0.95923447, 0.21798656, -0.02158783, 0.62183717,
-            0.04007379, 0.46044942, -0.10129622, 0.27720413, -0.37107886
-        ]).reshape((-1, 1))
-
-        indeps.add_output('x', x_vals)
-        indeps.add_output('y', np.array([
-            0.52577864, 0.30894559, 0.8420792, 0.35039912, -0.67290778,
-            -0.86236787, -0.97500023, 0.47739414, 0.51174103, 0.10052582
-        ]))
-        indeps.add_output('r', .7)
-
-        model.add_subsystem('circle', om.ExecComp('area = pi * r**2'))
-
-        model.add_subsystem('r_con', om.ExecComp('g = x**2 + y**2 - r**2',
-                                                 g=np.ones(SIZE), x=np.ones(SIZE), y=np.ones(SIZE)))
-
-        thetas = np.linspace(0, np.pi/4, SIZE)
-
-        model.add_subsystem('theta_con', om.ExecComp('g=arctan(y/x) - theta',
-                                                     g=np.ones(SIZE), x=np.ones(SIZE),
-                                                     y=np.ones(SIZE), theta=thetas))
-        model.add_subsystem('delta_theta_con', om.ExecComp('g = arctan(y/x)[::2]-arctan(y/x)[1::2]',
-                                                           g=np.ones(SIZE//2), x=np.ones(SIZE),
-                                                           y=np.ones(SIZE)))
-
-        model.add_subsystem('l_conx', om.ExecComp('g=x-1', g=np.ones(SIZE), x=np.ones(SIZE)))
-
-        model.connect('r', ('circle.r', 'r_con.r'))
-        model.connect('x', ['r_con.x', 'theta_con.x', 'delta_theta_con.x'])
-        model.connect('x', 'l_conx.x')
-        model.connect('y', ['r_con.y', 'theta_con.y', 'delta_theta_con.y'])
-
-        model.add_design_var('x', indices=[0, 3], flat_indices=True)
-        model.add_design_var('y')
-        model.add_design_var('r', lower=.5, upper=10)
-
-        # nonlinear constraints
-        model.add_constraint('r_con.g', equals=0)
-
-        IND = np.arange(SIZE, dtype=int)
-        EVEN_IND = IND[0::2]  # all even indices
-        model.add_constraint('theta_con.g', lower=-1e-5, upper=1e-5, indices=EVEN_IND)
-        model.add_constraint('delta_theta_con.g', lower=-1e-5, upper=1e-5)
-
-        # this constrains x[0] to be 1 (see definition of l_conx)
-        model.add_constraint('l_conx.g', equals=0, linear=False, indices=[0, ])
-
-        # linear constraint
-        model.add_constraint('y', equals=0, indices=[0], linear=True)
-
-        model.add_objective('circle.area', ref=-1)
-
-        prob.setup(mode='fwd')
-        prob.run_driver()
-        prob.cleanup()
-
-        # get the case we recorded
-        cr = om.CaseReader(self.filename)
-        case = cr.get_case(0)
-
-        # check 'use_indices' option, default is to use indices
-        dvs = case.get_design_vars()
-        assert_near_equal(dvs['x'], x_vals[[0, 3]], 1e-12)
-
-        dvs = case.get_design_vars(use_indices=False)
-        assert_near_equal(dvs['x'], x_vals, 1e-12)
-
-        cons = case.get_constraints()
-        self.assertEqual(len(cons['theta_con.g']), len(EVEN_IND))
-
-        cons = case.get_constraints(use_indices=False)
-        self.assertEqual(len(cons['theta_con.g']), SIZE)
-
-        # add one to all the inputs just to change the model, so we
-        # can see if loading the case values really changes the model
-        for name in prob.model._inputs:
-            model._inputs[name] += 1.0
-        for name in prob.model._outputs:
-            model._outputs[name] += 1.0
-
-        # load in the case we recorded and check that the model then matches
-        prob.load_case(case)
-        _assert_model_matches_case(case, model)
-
-    def test_multidimensional_arrays(self):
-        prob = om.Problem()
-        model = prob.model
-
-        comp = TestExplCompArray(thickness=1.)  # has 2D arrays as inputs and outputs
-        model.add_subsystem('comp', comp, promotes=['*'])
-        # just to add a connection, otherwise an exception is thrown in recording viewer data.
-        # must be a bug
-        model.add_subsystem('double_area',
-                            om.ExecComp('double_area = 2 * areas',
-                                        areas=np.zeros((2, 2)),
-                                        double_area=np.zeros((2, 2))),
-                            promotes=['*'])
-
-        prob.driver.add_recorder(self.recorder)
-        prob.driver.recording_options['includes'] = ['*']
-
-        prob.setup()
-        prob.run_driver()
-        prob.cleanup()
-
-        # Add one to all the inputs just to change the model
-        #   so we can see if loading the case values really changes the model
-        for name in prob.model._inputs:
-            model._inputs[name] += 1.0
-        for name in prob.model._outputs:
-            model._outputs[name] += 1.0
-
-        # Now load in the case we recorded
-        cr = om.CaseReader(self.filename)
-
-        driver_cases = cr.list_cases('driver', out_stream=None)
-        case = cr.get_case(driver_cases[0])
-
-        prob.load_case(case)
-
-        _assert_model_matches_case(case, model)
 
     def test_simple_paraboloid_scaled_desvars(self):
 
@@ -4213,28 +3785,6 @@ class TestPromAbsDict(unittest.TestCase):
         ]))
 
 
-def _assert_model_matches_case(case, system):
-    """
-    Check to see if the values in the case match those in the model.
-
-    Parameters
-    ----------
-    case: Case object
-        Case to be used for the comparison.
-    system: System object
-        System to be used for the comparison.
-    """
-    case_inputs = case.inputs
-    model_inputs = system._inputs
-    for name, model_input in model_inputs._abs_item_iter(flat=False):
-        np.testing.assert_almost_equal(case_inputs[name], model_input)
-
-    case_outputs = case.outputs
-    model_outputs = system._outputs
-    for name, model_output in model_outputs._abs_item_iter(flat=False):
-        np.testing.assert_almost_equal(case_outputs[name], model_output)
-
-
 @use_tempdirs
 class TestSqliteCaseReaderLegacy(unittest.TestCase):
 
@@ -4738,7 +4288,7 @@ class TestSqliteCaseReaderLegacy(unittest.TestCase):
         # Now load in the case we recorded
         prob.load_case(seventh_slsqp_iteration_case)
 
-        _assert_model_matches_case(seventh_slsqp_iteration_case, prob.model)
+        assert_model_matches_case(seventh_slsqp_iteration_case, prob.model)
 
     def test_driver_v2(self):
         """ Backwards compatibility version 2. """
@@ -4790,7 +4340,7 @@ class TestSqliteCaseReaderLegacy(unittest.TestCase):
         # Now load in the case we recorded
         prob.load_case(seventh_slsqp_iteration_case)
 
-        _assert_model_matches_case(seventh_slsqp_iteration_case, prob.model)
+        assert_model_matches_case(seventh_slsqp_iteration_case, prob.model)
 
     def test_solver_v2(self):
         """ Backwards compatibility version 2. """
@@ -4895,7 +4445,7 @@ class TestSqliteCaseReaderLegacy(unittest.TestCase):
         # Now load in the case we recorded
         prob.load_case(seventh_slsqp_iteration_case)
 
-        _assert_model_matches_case(seventh_slsqp_iteration_case, prob.model)
+        assert_model_matches_case(seventh_slsqp_iteration_case, prob.model)
 
     def test_driver_v1_pre_problem(self):
         """ Backwards compatibility oldest version. """
@@ -4942,62 +4492,8 @@ class TestSqliteCaseReaderLegacy(unittest.TestCase):
         # Now load in the case we recorded
         prob.load_case(seventh_slsqp_iteration_case)
 
-        _assert_model_matches_case(seventh_slsqp_iteration_case, prob.model)
+        assert_model_matches_case(seventh_slsqp_iteration_case, prob.model)
 
-
-class Adder(om.ExplicitComponent):
-    def setup(self):
-        self.add_input('x', shape_by_conn=True, distributed=True)
-        self.add_input('y')
-        self.add_output('x_sum', shape=1)
-        self.add_output('out_dist', copy_shape='x', distributed=True)
-
-    def compute(self, inputs, outputs):
-        outputs['x_sum'] = np.sum(inputs['x']) + inputs['y']**2.0
-        outputs['out_dist'] = inputs['x'] + 1.
-
-
-@unittest.skipUnless(MPI and PETScVector, "MPI and PETSc are required.")
-@use_tempdirs
-class SqliteCaseReaderMPI(unittest.TestCase):
-    N_PROCS = 2
-    def test_distrib_var_load(self):
-        prob = om.Problem()
-        ivc = prob.model.add_subsystem('ivc',om.IndepVarComp(), promotes=['*'])
-        ivc.add_output('x', val = np.ones(100 if prob.comm.rank == 0 else 10), distributed=True)
-        ivc.add_output('y', val = 1.0)
-
-        prob.model.add_subsystem('adder', Adder(), promotes=['*'])
-        prob.driver = om.ScipyOptimizeDriver(optimizer='SLSQP', tol=1e-9)
-        prob.model.add_design_var('y', lower=-1, upper=1)
-        prob.model.add_objective('x_sum')
-
-        recorder_file = 'distributed.sql'
-        prob.driver.add_recorder(om.SqliteRecorder(recorder_file))
-        prob.driver.recording_options['includes'] = ['*']
-        prob.driver.recording_options['record_objectives'] = True
-        prob.driver.recording_options['record_constraints'] = True
-        prob.driver.recording_options['record_desvars'] = True
-
-        prob.setup()
-        prob.run_driver()
-
-        reader = om.CaseReader(recorder_file)
-        # print(reader.list_cases())
-        prob.load_case(reader.get_case(-1))
-
-        with multi_proc_exception_check(prob.comm):
-            val = prob.get_val('adder.x', get_remote=False)
-            if prob.comm.rank == 0:
-                assert_near_equal(val, np.ones(100, dtype=float))
-            else:
-                assert_near_equal(val, np.ones(10, dtype=float))
-
-            val = prob.get_val('adder.out_dist', get_remote=False)
-            if prob.comm.rank == 0:
-                assert_near_equal(val, np.ones(100, dtype=float) + 1.)
-            else:
-                assert_near_equal(val, np.ones(10, dtype=float) + 1.)
 
 if __name__ == "__main__":
     unittest.main()
